refactor: resolve RPC rate limiting
